requirements.txt
Django==5.2.4
djangorestframework==3.16.0
django-crispy-forms==2.4
django-filter==25.1
django-tailwind==4.2.0
Pillow==11.3.0
mysqlclient==2.2.4
requests==2.32.4
python-decouple==3.8
django-allauth==0.60.1
django-cors-headers==4.3.1
